requirements.txt
Django==3.2.10
djangorestframework==3.12.4
djangorestframework_simplejwt==5.0.0
django-cors-headers==3.10.0
celery==5.2.3
requests==2.26.0
redis
